+++ lending/RELEASES.md
@@ -9,3 +9,13 @@ Based on:
 - [typescript v0.1.0] lending
 ### Releases
 - [NPM v0.1.0] https://www.npmjs.com/package/openapi/v/0.1.0 - lending
+
+## 2023-10-23 01:52:48
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.104.0 (2.169.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v0.2.0] lending
+### Releases
+- [NPM v0.2.0] https://www.npmjs.com/package/lending/v/0.2.0 - lending